delete PetType is already

+++ MyVet.Web/Controllers/PetTypesController.cs
@@ -125,25 +125,45 @@ namespace MyVet.Web.Controllers
             }
 
             var petType = await _context.PetTypes
+                .Include(pt => pt.Pets)
                 .FirstOrDefaultAsync(m => m.Id == id);
+           
             if (petType == null)
             {
                 return NotFound();
             }
 
-            return View(petType);
+            if (petType.Pets.Count > 0)
+            {
+                ModelState.AddModelError(string.Empty, "The pet Type cannot be removed.");
+                return RedirectToAction(nameof(Index));
+
+            }
+
+            try
+            {
+                _context.PetTypes.Remove(petType);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (Exception ex)
+            {
+
+                return BadRequest(ex.Message);
+
+            }
         }
 
         // POST: PetTypes/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
-        {
-            var petType = await _context.PetTypes.FindAsync(id);
-            _context.PetTypes.Remove(petType);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
-        }
+        //public async Task<IActionResult> DeleteConfirmed(int id)
+        //{
+        //    var petType = await _context.PetTypes.FindAsync(id);
+        //    _context.PetTypes.Remove(petType);
+        //    await _context.SaveChangesAsync();
+        //    return RedirectToAction(nameof(Index));
+        //}
 
         private bool PetTypeExists(int id)
         {
